backend/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Database
psycopg2-binary>=2.9,<3.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0
PyJWT>=2.8,<3.0
bcrypt>=4.1,<5.0

# Media Handling
Pillow>=10.0,<11.0
django-imagekit>=4.1,<5.0

# Utilities
python-dotenv>=1.0,<2.0

# Django Extensions
django-extensions>=3.2,<4.0
